NLP - Commentator/research/10_llm_commentary/requirements.txt
# Requirements for Phase 10: LLM Commentary Generation

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0

# Progress bars (optional)
tqdm>=4.65.0

# Comparison metrics (from Phase 8)
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

